requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==24.1.0
aiogram==2.15
aiohttp==3.9.5
aiohttp_retry==2.9.1
aiormq==6.4.2
aiosignal==1.3.2
APScheduler==3.9.1.post1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
babel==2.17.0
bidict==0.23.1
blinker==1.9.0
boto3==1.38.39
botocore==1.38.39
CacheControl==0.12.14
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
colorzero==2.0
confluent-kafka==2.10.1
cryptography==45.0.4
cycler==0.12.1
dask==2022.10.2
dnspython==2.3.0
docopt==0.6.2
fbmessenger==6.0.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==20181003210633
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpiozero==2.0.1
greenlet==3.2.3
grpcio==1.73.0
gTTS==2.5.4
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.4
httptools==0.6.4
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonpickle==3.0.4
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.8
libclang==18.1.1
locket==1.0.0
Markdown==3.8.1
MarkupSafe==3.0.2
matplotlib==3.5.3
mattermostwrapper==2.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.1
multidict==5.2.0
networkx==2.6.3
numpy==1.23.5
oauthlib==3.3.0
opt_einsum==3.4.0
packaging==20.9
pamqp==3.2.1
partd==1.4.2
pigpio==1.78
pillow==11.2.1
pluggy==1.6.0
portalocker==2.10.1
prompt-toolkit==3.0.28
propcache==0.3.2
protobuf==4.23.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==1.10.9
pydot==1.4.2
PyJWT==2.10.1
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.2.3
pyrsistent==0.20.0
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2022.7.1
PyYAML==6.0.2
questionary==1.10.0
randomname==0.1.5
rasa==3.6.21
rasa-sdk==3.6.2
redis==4.6.0
regex==2022.10.31
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rocketchat-API==1.30.0
RPi.GPIO==0.7.1
rsa==4.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.12
s3transfer==0.13.0
safetensors==0.4.5
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.10.1
sentry-sdk==1.14.0
simple-websocket==1.1.0
six==1.17.0
sklearn-crfsuite==0.3.6
skops==0.9.0
slack_sdk==3.35.0
SQLAlchemy==1.4.54
structlog==23.3.0
structlog-sentry==2.1.0
sympy==1.14.0
tabulate==0.9.0
tarsafe==0.0.4
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-cpu-aws==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==3.1.0
terminaltables==3.1.10
threadpoolctl==3.6.0
toolz==1.0.0
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
twilio==8.2.2
typing-utils==0.1.0
typing_extensions==4.14.0
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.5.0
uvloop==0.21.0
wcwidth==0.2.13
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==3.1.3
wrapt==1.14.1
wsproto==1.2.0
yarl==1.20.1
